Updated testSumOddsUint8 (to use correct data types for returned values)

+++ assignment11Demo/assignment11Demo.ino
@@ -131,8 +131,8 @@ void testSumOddsUint8() {
   for(int i=0;i<numTests;i++) {
     uint8_t arg1 = arg1s[i];
     uint8_t arg2 = arg2s[i];
-    uint8_t algorithm = sumOdds(arg1, arg2);
-    uint8_t assembly = sumOddsUint8(arg1, arg2);
+    uint16_t algorithm = sumOdds(arg1, arg2);
+    uint16_t assembly = sumOddsUint8(arg1, arg2);
     uint8_t expected = results[i];
     pprintf("sumOdds(%u,%u) is %u (algorithm) and %u (assembly); Should be %u", arg1, arg2, algorithm, assembly, expected);
     // Compare the results of the algorithm and real answer
